Include dependencies for main entry point
